refactor: rename SchemaTranslator and SchemaValidator to Translator and Validator
This commit renames the `SchemaTranslator` and `SchemaValidator` classes to `Translator` and `Validator` respectively, along with updating all related imports and documentation. The change aims to simplify class names and improve code readability without altering functionality.
