+++ docs/source/quickstart.rst
@@ -9,17 +9,17 @@ Basic Usage
 1. Define a Schema
 ~~~~~~~~~~~~~~~~~~
 
-First, define your schema using belso's Schema and BaseField classes:
+First, define your schema using belso's Schema and Field classes:
 
 .. code-block:: python
 
-   from belso.schemas import Schema, BaseField
+   from belso import Schema, Field
 
    class UserSchema(Schema):
        name = "UserSchema"
        fields = [
-           BaseField(name="name", type=str, description="User's name", required=True),
-           BaseField(name="age", type=int, description="User's age", required=True)
+           Field(name="name", type_hint=str, description="User's name", required=True),
+           Field(name="age", type_hint=int, description="User's age", required=True)
        ]
 
 2. Translate to Provider Format
@@ -29,14 +29,14 @@ Convert your schema to a specific provider format:
 
 .. code-block:: python
 
+    from belso import Translator
     from belso.utils import PROVIDERS
-    from belso.translator import SchemaTranslator
 
     # Convert to OpenAI format
-    openai_schema = SchemaTranslator.translate(UserSchema, to=PROVIDERS.OPENAI)
+    openai_schema = Translator.translate(UserSchema, to=PROVIDERS.OPENAI)
 
     # Convert to Anthropic format
-    anthropic_schema = SchemaTranslator.translate(UserSchema, to=PROVIDERS.ANTHROPIC)
+    anthropic_schema = Translator.translate(UserSchema, to=PROVIDERS.ANTHROPIC)
 
 3. Validate Data
 ~~~~~~~~~~~~~~~~
@@ -45,7 +45,7 @@ Validate data against your schema:
 
 .. code-block:: python
 
-   from belso.validator import SchemaValidator
+   from belso import Validator
 
    data = {"name": "John", "age": 30}
-   validated_data = SchemaValidator.validate(data, UserSchema)
+   validated_data = Validator.validate(data, UserSchema)
